Modify DebtRejection code

+++ etc/databases/unpaid.sql
@@ -11,6 +11,6 @@ CREATE TABLE `debt_rejections` (
     `debtor_account` VARCHAR(30) NOT NULL,
     `creditor_account` VARCHAR(30) NOT NULL,
     `debtor_name` VARCHAR(255) NOT NULL,
-    `process_status` TINYINT(1) NOT NULL,
+    `process_status` SMALLINT NOT NULL,
     PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
